backend/requirements.txt
fastapi
uvicorn
pydantic
PyGithub
python-multipart
python-dotenv
requests
